remove custom feed harness from appview v2
