--- packages/bsky/src/api/feed-gen/index.ts
@@ -1,6 +1,0 @@
-import { MountedAlgos } from './types'
-
-// These are custom algorithms that will be mounted directly onto an AppView
-// Feel free to remove, update to your own, or serve the following logic at a record that you control
-// @TODO currently empty during appview v2 build out
-export const makeAlgos = (_did: string): MountedAlgos => ({})
