+++ packages/bsky/src/api/index.ts
@@ -1,6 +1,5 @@
 import { Server } from '../lexicon'
 import AppContext from '../context'
-import describeFeedGenerator from './app/bsky/feed/describeFeedGenerator'
 import getTimeline from './app/bsky/feed/getTimeline'
 import getActorFeeds from './app/bsky/feed/getActorFeeds'
 import getSuggestedFeeds from './app/bsky/feed/getSuggestedFeeds'
@@ -8,7 +7,6 @@ import getAuthorFeed from './app/bsky/feed/getAuthorFeed'
 import getFeed from './app/bsky/feed/getFeed'
 import getFeedGenerator from './app/bsky/feed/getFeedGenerator'
 import getFeedGenerators from './app/bsky/feed/getFeedGenerators'
-import getFeedSkeleton from './app/bsky/feed/getFeedSkeleton'
 import getLikes from './app/bsky/feed/getLikes'
 import getListFeed from './app/bsky/feed/getListFeed'
 import getPostThread from './app/bsky/feed/getPostThread'
@@ -40,7 +38,6 @@ import listNotifications from './app/bsky/notification/listNotifications'
 import updateSeen from './app/bsky/notification/updateSeen'
 import registerPush from './app/bsky/notification/registerPush'
 import getPopularFeedGenerators from './app/bsky/unspecced/getPopularFeedGenerators'
-import getTimelineSkeleton from './app/bsky/unspecced/getTimelineSkeleton'
 import getTaggedSuggestions from './app/bsky/unspecced/getTaggedSuggestions'
 import getSubjectStatus from './com/atproto/admin/getSubjectStatus'
 import updateSubjectStatus from './com/atproto/admin/updateSubjectStatus'
@@ -57,7 +54,6 @@ export * as blobResolver from './blob-resolver'
 
 export default function (server: Server, ctx: AppContext) {
   // app.bsky
-  describeFeedGenerator(server, ctx)
   getTimeline(server, ctx)
   getActorFeeds(server, ctx)
   getSuggestedFeeds(server, ctx)
@@ -65,7 +61,6 @@ export default function (server: Server, ctx: AppContext) {
   getFeed(server, ctx)
   getFeedGenerator(server, ctx)
   getFeedGenerators(server, ctx)
-  getFeedSkeleton(server, ctx)
   getLikes(server, ctx)
   getListFeed(server, ctx)
   getPostThread(server, ctx)
@@ -97,7 +92,6 @@ export default function (server: Server, ctx: AppContext) {
   updateSeen(server, ctx)
   registerPush(server, ctx)
   getPopularFeedGenerators(server, ctx)
-  getTimelineSkeleton(server, ctx)
   getTaggedSuggestions(server, ctx)
   // com.atproto
   getSubjectStatus(server, ctx)
