+++ packages/pds/tests/proxied/feedgen.test.ts
@@ -1,70 +1,55 @@
-import AtpAgent, { AtUri, AppBskyFeedNS } from '@atproto/api'
+import AtpAgent, { AtUri } from '@atproto/api'
 import { TestNetwork, SeedClient } from '@atproto/dev-env'
 import basicSeed from '../seeds/basic'
 import { forSnapshot } from '../_util'
+import { InvalidRequestError } from '@atproto/xrpc-server'
 
 describe('feedgen proxy view', () => {
   let network: TestNetwork
   let agent: AtpAgent
   let sc: SeedClient
-
-  const origGetFeedGenerator = AppBskyFeedNS.prototype.getFeedGenerator
-  const feedUri = AtUri.make(
-    'did:example:feed-publisher',
-    'app.bsky.feed.generator',
-    'mutuals',
-  )
+  let feedUri: AtUri
 
   beforeAll(async () => {
     network = await TestNetwork.create({
       dbPostgresSchema: 'proxy_feedgen',
-      bsky: {
-        // @TODO consider using makeAlgos() here if the appview begins supporting any feeds out of the box
-        algos: {
-          [feedUri.toString()]: async () => ({ feedItems: [] }),
-        },
-      },
     })
     agent = network.pds.getClient()
     sc = network.getSeedClient()
     await basicSeed(sc, { addModLabels: network.bsky })
+    feedUri = AtUri.make(sc.dids.alice, 'app.bsky.feed.generator', 'mutuals')
+
+    const feedGen = await network.createFeedGen({
+      [feedUri.toString()]: ({ params }) => {
+        if (params.feed !== feedUri.toString()) {
+          throw new InvalidRequestError('Unknown feed')
+        }
+        return {
+          encoding: 'application/json',
+          body: {
+            feed: [
+              { post: sc.posts[sc.dids.alice][0].ref.uriStr },
+              { post: sc.posts[sc.dids.carol][0].ref.uriStr },
+            ],
+          },
+        }
+      },
+    })
+
     // publish feed
-    const feed = await agent.api.app.bsky.feed.generator.create(
+    await agent.api.app.bsky.feed.generator.create(
       { repo: sc.dids.alice, rkey: feedUri.rkey },
       {
-        did: network.bsky.ctx.cfg.feedGenDid ?? '',
-        displayName: 'Mutuals',
+        did: feedGen.did,
+        displayName: 'Test feed',
         createdAt: new Date().toISOString(),
       },
       sc.getHeaders(sc.dids.alice),
     )
     await network.processAll()
-    // mock getFeedGenerator() for use by pds's getFeed since we don't have a proper feedGenDid or feed publisher
-    AppBskyFeedNS.prototype.getFeedGenerator = async function (params, opts) {
-      if (params?.feed === feedUri.toString()) {
-        return {
-          success: true,
-          data: {
-            isOnline: true,
-            isValid: true,
-            view: {
-              cid: feed.cid,
-              uri: feed.uri,
-              did: network.bsky.ctx.cfg.feedGenDid ?? '',
-              creator: { did: sc.dids.alice, handle: 'alice.test' },
-              displayName: 'Mutuals',
-              indexedAt: new Date().toISOString(),
-            },
-          },
-          headers: {},
-        }
-      }
-      return origGetFeedGenerator.call(this, params, opts)
-    }
   })
 
   afterAll(async () => {
-    AppBskyFeedNS.prototype.getFeedGenerator = origGetFeedGenerator
     await network.close()
   })
